TEC-7379 Update typo from UCG to UGC
