+++ src/test/java/com/echobox/api/linkedin/types/ugc/UGCShareTest.java
@@ -24,11 +24,6 @@ import com.echobox.api.linkedin.jsonmapper.DefaultJsonMapper;
 import com.echobox.api.linkedin.jsonmapper.DefaultJsonMapperTestBase;
 import com.echobox.api.linkedin.types.objectype.AuditStamp;
 import com.echobox.api.linkedin.types.social.actions.CommentAction;
-import com.echobox.api.linkedin.types.ucg.Commentary;
-import com.echobox.api.linkedin.types.ucg.ShareContent;
-import com.echobox.api.linkedin.types.ucg.ShareMedia;
-import com.echobox.api.linkedin.types.ucg.TargetAudience;
-import com.echobox.api.linkedin.types.ucg.UGCShare;
 import com.echobox.api.linkedin.types.urn.URN;
 
 import org.junit.Test;
